Many things
Setting a working CMakeFile

Setting #includes as required

Bug fixing

+++ src/server0/upload.cpp
@@ -5,7 +5,7 @@ unsigned int getFileLength( std::string const& filename ) {						//Retrieving fi
 	std::ifstream file( filename.c_str(), std::ios::binary | std::ios::in );
 	if( !file.fail() ){
 		file.seekg( 0, std::ios::end );
-		return file.tellg();
+		return static_cast<unsigned int>( file.tellg() );
 	}
 	return 0;
 }
